Changed license from CC-BY-SA 4.0 to Apache 2.0
Originally, the idea was to use Creative Commons Attribution-ShareAlike 4.0 International as license for this repository (which is currently still empty ;-) ). That license has a really nice section on attribution. Unfortunately, CC-BY-SA is not a good license for software projects (see also: https://creativecommons.org/faq/#can-i-apply-a-creative-commons-license-to-software), so before adding any other files, we have changed the license to Apache 2.0.