Fixed Bug that caused wrong timing with different tempi
When copying events from one phrase to another one that has a different tempo, the timing of the events ended up being wrong. The reason was that for quantization, instead of using the target phrase, the current phrase was being used. So we were quite lucky to have found this one now and not later (because it also reproduced only under specific circumstances)